feat: move stats to own file
